+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Format based on [Keep a Changelog](https://keepachangelog.com/), adapted for int
 
 ## [Unreleased]
 
+## [523] - 2026-03-25
+
+- [TASK-31] Add topics column to conventions and tusk conventions search command
+
 ## [522] - 2026-03-25
 
 - [TASK-30] Add Step 1b to /tusk-update: migrate CLAUDE.md conventions to DB